# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.DoctorLoginReq:
    properties:
      phone_number:
        type: string
    type: object
  models.DoctorLoginRes:
    properties:
      access_token:
        type: string
      clinic_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  models.DoctorReqForUI:
    properties:
      clinic_id:
        type: integer
      end_time:
        type: string
      experience:
        type: integer
      first_name:
        type: string
      floor:
        type: integer
      image_link:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      room_number:
        type: integer
      service_id:
        type: integer
      start_time:
        type: string
      work_day:
        type: string
    type: object
  models.DoctorResInfo:
    properties:
      clinic_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      image_link:
        type: string
      last_name:
        type: string
      service_id:
        type: integer
      work_day:
        type: string
    type: object
  models.DoctorResponse:
    properties:
      clinic_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      experience:
        type: integer
      first_name:
        type: string
      floor:
        type: integer
      id:
        type: integer
      image_link:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      room_number:
        type: integer
      service_id:
        type: integer
      service_name:
        type: string
      start_time:
        type: string
      work_day:
        type: string
    type: object
  models.Doctors:
    properties:
      doctors:
        items:
          $ref: '#/definitions/models.DoctorResInfo'
        type: array
    type: object
  models.DoctorsList:
    properties:
      doctors:
        items:
          $ref: '#/definitions/models.DoctorResponse'
        type: array
    type: object
...
  models.GetCustomersOfDoctor:
    properties:
      client_id:
        type: integer
      doctor_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      status:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  models.UpdateDoctor:
    properties:
      end_time:
        type: string
      experience:
        type: integer
      first_name:
        type: string
      floor:
        type: integer
      id:
        type: integer
      image_link:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      room_number:
        type: integer
      start_time:
        type: string
      work_day:
        type: string
    type: object
...
info:
  contact: {}
  description: This is Clinic server api
  termsOfService: 1 term Traffic Light
  title: Clinic
  version: "1.0"
paths:
...
  /doctor:
    post:
      consumes:
      - application/json
      description: Create Doctor
      operationId: createdoctor
      parameters:
      - description: CreateAdminInfo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.DoctorReqForUI'
      produces:
      - application/json
      responses:
        "201":
          description: Admin data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DoctorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Doctor
      tags:
      - Doctor
  /doctor/{id}:
    get:
      consumes:
      - application/json
      description: Get Doctor
      operationId: getdoctorbyid
      parameters:
      - description: doctor_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DoctorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Doctor
      tags:
      - Doctor
  /doctor/customers/{doctor_id}:
    get:
      consumes:
      - application/json
      operationId: getcustomersbydoctorid
      parameters:
      - description: doctor_id
        in: path
        name: doctor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: customer data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.GetCustomersOfDoctor'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Doctor's customers
      tags:
      - Doctor
  /doctor/login:
    post:
      consumes:
      - application/json
      description: Login Doctor
      operationId: logindoctor
      parameters:
      - description: DoctorLogin
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.DoctorLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DoctorLoginRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login Doctor
      tags:
      - LOGIN
  /doctor/service/{clinic_id}:
    get:
      consumes:
      - application/json
      description: Get Doctor by ServiceName
      operationId: getdoctorsbyservicename
      parameters:
      - description: clinicID
        in: path
        name: clinic_id
        required: true
        type: integer
      - description: ServiceName
        in: query
        name: servicename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Doctors'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Doctor by ServiceName
      tags:
      - Doctor
  /doctors:
    put:
      consumes:
      - application/json
      description: PUT Doctors
      operationId: putdoctors
      parameters:
      - description: UpdateClient
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDoctor'
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Doctors
      tags:
      - Doctor
  /doctors/{clinic_id}:
    get:
      consumes:
      - application/json
      description: Get all  Doctors List
      operationId: getlistdoctor
      parameters:
      - description: ClinicID
        in: path
        name: clinic_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DoctorsList'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all Doctor
      tags:
      - Doctor
  /doctors/delete/{id}:
    delete:
      consumes:
      - application/json
      description: DELETE Doctors
      operationId: deletedoctor
      parameters:
      - description: doctorId
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete Doctors
      tags:
      - Doctor
  /doctors/search/{clinic_id}:
    get:
      consumes:
      - application/json
      description: Get all  Doctor
      operationId: getalldoctor
      parameters:
      - description: ClinicID
        in: path
        name: clinic_id
        required: true
        type: integer
      - description: keyword
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DoctorsList'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all Doctor
      tags:
      - Doctor
  /doctors/service/{service_id}:
    get:
      consumes:
      - application/json
      description: Get Doctor by ServiceID
      operationId: getdoctorsbyserviceid
      parameters:
      - description: serviceID
        in: path
        name: service_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DoctorsList'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Doctor by ServiceID
      tags:
      - Doctor
...
  /service/delete/{id}:
    delete:
      consumes:
      - application/json
      description: PUT Service
      operationId: service
      parameters:
      - description: serviceID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Service data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Service
      tags:
      - Service
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"